moved create methods into business logic
