+++ eq-author-api/src/businessLogic/createQuestionPage.js
@@ -1,0 +1,18 @@
+const { v4: uuidv4 } = require("uuid");
+
+const createQuestionPage = (input = {}) => ({
+  id: uuidv4(),
+  pageType: "QuestionPage",
+  title: "",
+  description: "",
+  descriptionEnabled: false,
+  guidanceEnabled: false,
+  definitionEnabled: false,
+  additionalInfoEnabled: false,
+  answers: [],
+  routing: null,
+  alias: null,
+  ...input,
+});
+
+module.exports = createQuestionPage;
