+++ eq-author-api/src/businessLogic/index.js
@@ -6,6 +6,10 @@ const createExpression = require("./createExpression");
 const createLeftSide = require("./createLeftSide");
 const getNextDestination = require("./getNextDestination");
 const createRightSide = require("./createRightSide");
+const createQuestionPage = require("./createQuestionPage");
+const createCalculatedSummary = require("./createCalculatedSummary");
+const createFolder = require("./createFolder");
+const createSection = require("./createSection");
 
 module.exports = {
   createRouting,
@@ -16,4 +20,8 @@ module.exports = {
   createLeftSide,
   getNextDestination,
   createRightSide,
+  createQuestionPage,
+  createCalculatedSummary,
+  createFolder,
+  createSection,
 };
